Moved composition to sub package
