+++ tardis/model/base.py
@@ -6,7 +6,6 @@ from astropy import units as u
 from tardis import constants
 import radioactivedecay as rd
 from radioactivedecay.utils import Z_DICT
-from tardis.model.geometry.radial1d import Radial1DGeometry
 
 from tardis.util.base import quantity_linspace, is_valid_nuclide_or_elem
 from tardis.io.parsers.csvy import load_csvy
@@ -21,61 +20,10 @@ from tardis.io.config_reader import Configuration
 from tardis.io.util import HDFWriterMixin
 from tardis.io.decay import IsotopeAbundances
 from tardis.model.density import HomologousDensity
+from tardis.model.geometry.radial1d import Radial1DGeometry
+from tardis.model.composition.composition import Composition
 
 logger = logging.getLogger(__name__)
-
-
-class Composition:
-    """
-    Holds information about model composition
-
-    Parameters
-    ----------
-    density : astropy.units.quantity.Quantity
-        An array of densities for each shell.
-    isotopic_mass_fraction : pd.DataFrame
-    atomic_mass : pd.DataFrame
-    atomic_mass_unit: astropy.units.Unit
-
-    Attributes
-    ----------
-    atomic_mass : pd.DataFrame
-        Atomic mass of elements calculated for each shell.
-    elemental_number_density : pd.DataFrame
-        Number density of each element in each shell.
-    """
-
-    def __init__(
-        self,
-        density,
-        elemental_mass_fraction,
-        atomic_mass,
-        atomic_mass_unit=u.g,
-    ):
-        self.density = density
-        self.elemental_mass_fraction = elemental_mass_fraction
-        self.atomic_mass_unit = atomic_mass_unit
-        self._atomic_mass = atomic_mass
-
-    @property
-    def atomic_mass(self):
-        """Atomic mass of elements in each shell"""
-        if self._atomic_mass is None:
-            raise AttributeError(
-                "ModelState was not provided elemental masses."
-            )
-        return self._atomic_mass
-
-    @property
-    def elemental_number_density(self):
-        """Elemental Number Density computed using the formula: (elemental_mass_fraction * density) / atomic mass"""
-        if self.atomic_mass is None:
-            raise AttributeError(
-                "ModelState was not provided elemental masses."
-            )
-        return (self.elemental_mass_fraction * self.density).divide(
-            self.atomic_mass, axis=0
-        )
 
 
 class ModelState:
